[9/n] rename daft-derive internals/mod.rs to imp.rs
Going to add more modules here -- I like not putting things in `mod.rs` because
private fields of structs in `mod.rs` can be accessed anywhere in the module.

This is pure code movement with no functional changes.

Reviewers: andrewjstone

Reviewed By: andrewjstone

Pull Request: https://github.com/oxidecomputer/daft/pull/40
